Fix potential infinite loop during Pattern application
The while loop was crashing my browser, so converted into a for loop. The issue is when appendChild does not remove the child from __defs.childNodes, I’m not entirely sure why this was happening, but this method also is slightly more optimized as you don’t recalculate the length upon each interation.